+++ crates/zizmor/src/models/uses.rs
@@ -1,10 +1,16 @@
 //! Extension traits for the `Uses` APIs.
+//!
+//! This is for GitHub Actions style `uses:` clauses;
+//! more general "reference to repository" handling
+//! lives in [`super::repo_ref`].
 
 use std::{str::FromStr, sync::LazyLock};
 
 use github_actions_models::common::{RepositoryUses, Uses};
 use regex::Regex;
 use serde::Deserialize;
+
+use crate::models::repo_ref::RepoRef;
 
 /// Matches all variants of [`RepositoryUsesPattern`] except `*`.
 ///
@@ -193,6 +199,8 @@ impl<'de> Deserialize<'de> for RepositoryUsesPattern {
 }
 
 /// Useful APIs for interacting with `uses: owner/repo` clauses.
+///
+/// Some of these APIs are projections of [`RepoRef`]'s APIs.
 pub(crate) trait RepositoryUsesExt {
     /// Returns whether this `uses:` clause matches the given pattern.
     ///
@@ -200,20 +208,11 @@ pub(crate) trait RepositoryUsesExt {
     /// same matching rules.
     fn matches(&self, pattern: &str) -> bool;
 
-    /// Returns whether this `uses:` clause has a `git` ref and, if so,
-    /// whether that ref is a commit ref.
-    ///
-    /// For example, `foo/bar@baz` returns false while `foo/bar@1234...`
-    /// returns true.
+    /// See [`RepoRef::ref_is_commit`].
     fn ref_is_commit(&self) -> bool;
 
-    /// Returns the `git` ref for this `uses:`, if present.
+    /// See [`RepoRef::commit_ref`].
     fn commit_ref(&self) -> Option<&str>;
-
-    /// Returns the *symbolic* `git` ref for this `uses`, if present.
-    ///
-    /// Commit refs (i.e. SHA refs) are not returned.
-    fn symbolic_ref(&self) -> Option<&str>;
 }
 
 impl RepositoryUsesExt for RepositoryUses {
@@ -226,21 +225,11 @@ impl RepositoryUsesExt for RepositoryUses {
     }
 
     fn ref_is_commit(&self) -> bool {
-        self.git_ref().len() == 40 && self.git_ref().chars().all(|c| c.is_ascii_hexdigit())
+        RepoRef::from(self).ref_is_commit()
     }
 
     fn commit_ref(&self) -> Option<&str> {
-        match &self.git_ref() {
-            git_ref if self.ref_is_commit() => Some(git_ref),
-            _ => None,
-        }
-    }
-
-    fn symbolic_ref(&self) -> Option<&str> {
-        match &self.git_ref() {
-            git_ref if !self.ref_is_commit() => Some(git_ref),
-            _ => None,
-        }
+        RepoRef::from(self).commit_ref()
     }
 }
 
@@ -270,7 +259,7 @@ impl UsesExt for Uses {
             // and the "hashedness" of a local action is mostly moot anyways
             // (since it's fully contained within the calling repo),
             Uses::Local(_) => false,
-            Uses::Repository(repo) => !repo.ref_is_commit(),
+            Uses::Repository(repo) => !RepoRef::from(repo).ref_is_commit(),
             Uses::Docker(docker) => docker.hash().is_none(),
         }
     }